Add simple model cache to allow simple, multiple context
